check size before copying the temporary fstab to /etc/fstab in nfsmount (Bug #23513)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs@30186 1b283449-9f2f-0410-b571-8eb228e1a901

+++ base/univention-base-files/univention-directory-policy/nfsmounts.py
@@ -215,7 +215,8 @@ def main():
 	f_new.close()
 	debug('Switching /etc/fstab...\n')
 	if not simulate:
-		os.rename(fstabNew, '/etc/fstab')
+		if os.path.isfile(fstabNew) and os.path.getsize(fstabNew) > 0:
+			os.rename(fstabNew, '/etc/fstab')
 
 	fp = open('/etc/mtab', 'r')
 	for line in fp:
